pyproject.toml
[project]
name = "ora-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "oracledb>=3.4.2",
    "python-dotenv>=1.2.2",
]
